Replace Configuration interface by StaticConfiguration class
Originally Configuration interface was introduced to implement different
loading strategies but after they was implemented via
ConfigurationUtils. Original interface is not necessary anymore as POJO
seems suit all purposes well.
