+++ README.md
@@ -86,7 +86,7 @@
 
 ```java
     // Create SDK configuration
-    StaticConfiguration config = StaticConfiguration.newBuilder()
+    Configuration config = Configuration.newBuilder()
         .setEthereumJsonRpcEndpoint("https://mainnet.infura.io/v3/" /* + your infura project id */)
         .setIpfsEndpoint("http://ipfs.singularitynet.io:80")
         .setIdentityType(Configuration.IdentityType.PRIVATE_KEY)
